test(e2e): harness-agent absorbs 12 standalone probes
Migrates 12 of the 19 originally-targeted standalone e2e probes into
harness-agent. The other 7 from the original list were already absorbed
by earlier PRs (esc-interrupt, msg-queue, streaming-journey-* in #321;
cwd-popover-recent-unfiltered, focus-orchestration in harness-ui #317)
or never existed under those names.

Absorbed (case id = original probe name with probe-e2e- prefix stripped):

Pure-store cases (no real CLI required):
- empty-group-new-session
- interrupt-banner
- tool-journey-render

Real-CLI cases (use #335 capability `requiresClaudeBin: true`, skip on
machines without claude on PATH):
- tool-call-dogfood
- input-queue
- send
- switch
- delete-session-kills-process
- default-cwd (also `userDataDir: 'fresh'` — tests cold-launch ~ expansion)
- streaming-partial-frames (also `userDataDir: 'fresh'` — wire-level capture
  needs clean IPC frame stream)

Heavy state-mutating cases (placed at end of cases array):
- notify-integration (mutates main-process notify bootstrap; placed last
  before close-window so its mutations don't leak to other cases)
- close-window-aborts-sessions (uses `relaunch: true` because emitting
  before-quit puts the app into is-quitting mode — fresh app required)

Layer 1 decisions:
- All 12 absorbable. None reclassified to bucket 4 (restore) — the cases
  needing fresh state use `userDataDir: 'fresh'` (one-off cold launch),
  not full restore semantics.
- 7 "missing" probes from the original 19-list are already absorbed
  elsewhere; called out in PR body. No new work needed.

Files deleted:
  scripts/probe-e2e-close-window-aborts-sessions.mjs
  scripts/probe-e2e-default-cwd.mjs
  scripts/probe-e2e-delete-session-kills-process.mjs
  scripts/probe-e2e-empty-group-new-session.mjs
  scripts/probe-e2e-input-queue.mjs
  scripts/probe-e2e-interrupt-banner.mjs
  scripts/probe-e2e-notify-integration.mjs
  scripts/probe-e2e-send.mjs
  scripts/probe-e2e-streaming-partial-frames.mjs
  scripts/probe-e2e-switch.mjs
  scripts/probe-e2e-tool-call-dogfood.mjs
  scripts/probe-e2e-tool-journey-render.mjs
